+++ texter.py
@@ -8,8 +8,5 @@ client = TwilioRestClient(cfg.TWILIO_ACCOUNT_SID, cfg.TWILIO_AUTH_TOKEN)
 
 message = client.messages.create(
     to="+18477574663",
-    from_=cfg.TWILIO_FROM_NUMBER,
+    from_="+13124706976",
     body="Hello, World!")
-
-print(message)
-print(message.sid)
